Move AudioNode trait to src/node/audio_node.rs
